Force caret-color in Forced Colors Mode
When Forced Colors happened at computed value time, caret-color would
pick up the forced color through currentcolor. However, now that
we are forcing colors at used value time, it appeared that caret-color
regressed since it no longer picks up the forced currentcolor.

This change adds caret-color to the properties that are adjusted in
Forced Colors Mode, and forces it to use the forced currentcolor.

I've opened a new spec issue to address this, as well as a few other
color properties that I noticed would also be affected by this:
https://github.com/w3c/csswg-drafts/issues/5873

Bug: 1166392
Change-Id: I239700441ed17bfd0b33443887a0c72793843cf1